apps/gnosi/requirements.txt
python-multipart>=0.0.9
PyYAML>=6.0
dotenv>=0.9.9
python-dotenv>=1.0.0
Flask>=3.0.0
flask-cors>=4.0.0
requests>=2.31.0
httpx>=0.27.0
notion-client>=2.0.0
networkx>=3.0
numpy>=1.24.0
gunicorn>=20.1.0
cloudinary>=1.36.0
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
langgraph>=0.0.10
langchain>=0.1.0
langchain-openai>=0.0.5
pydantic>=2.5.0
pytest>=7.0.0
chromadb>=0.4.0
langchain-chroma>=0.1.0
langgraph-checkpoint-sqlite>=1.0.0
psutil>=5.9.0
simpleeval>=0.9.13
sentence-transformers>=2.2.0
langchain-huggingface>=0.0.1
icalendar>=5.0.0
google-auth>=2.0.0
google-auth-oauthlib>=1.0.0
google-api-python-client>=2.0.0
groq>=0.5.0
anthropic>=0.20.0
cohere>=5.0.0
sqlalchemy>=2.0.0
pyzotero>=1.5.0
feedparser>=6.0.0
beautifulsoup4>=4.12.0
alembic>=1.13.0
gTTS>=2.5.0
httplib2>=0.22.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
cryptography>=42.0.0
aiosqlite>=0.20.0
Pillow>=10.0.0
